refactor: Implement build.prop reader
This implementation is cleaner than the old one, since it allows us to have all build properties extracted information in one class and provides an easy access to the required properties as class attributes. Also, the current implementation uses regex for extracting build properties instead of text slicing while looping through the the file lines It also handles the case of not found properties.

Signed-off-by: yshalsager <ysh-alsager@hotmail.com>